build ubuntu-25.10

+++ pkg/stages/steps_init.go
@@ -241,28 +241,35 @@ func GetWorkaroundsStage(sis values.System, l types.KairosLogger) []schema.Stage
 
 	if config.DefaultConfig.TrustedBoot {
 		if sis.Distro == values.Ubuntu {
-			kernel, err := getLatestKernel(l)
-			if err != nil {
-				l.Logger.Error().Msgf("Failed to get the latest kernel: %s", err)
-				return stages
+			systemVersion, err := semver.NewVersion(sis.Version)
+			if err == nil {
+				// 25.10 is the first version where this workaround is not needed
+				threshold, _ := semver.NewVersion("25.10")
+				if systemVersion.LessThan(threshold) {
+					kernel, err := getLatestKernel(l)
+					if err != nil {
+						l.Logger.Error().Msgf("Failed to get the latest kernel: %s", err)
+						return stages
+					}
+					// This looks like its out of its place as we would expect this modules to be in the initrd but this is for Trusted Boot
+					// so the initrd is creating during artifact build and contains the rootfs, so this is ok to be in here
+					stages = append(stages, []schema.Stage{
+						{
+							Name:     "Download linux-modules-extra for nvdimm modules",
+							OnlyIfOs: "Ubuntu.*",
+							Commands: []string{
+								fmt.Sprintf("apt-get download linux-modules-extra-%s", kernel),
+								fmt.Sprintf("dpkg-deb -x linux-modules-extra-%s_*.deb /tmp/modules", kernel),
+								fmt.Sprintf("mkdir -p /usr/lib/modules/%s/kernel/drivers/nvdimm", kernel),
+								fmt.Sprintf("mv /tmp/modules/lib/modules/%[1]s/kernel/drivers/nvdimm/* /usr/lib/modules/%[1]s/kernel/drivers/nvdimm/", kernel),
+								fmt.Sprintf("depmod -a %s", kernel),
+								"rm -rf /tmp/modules",
+								"rm /*.deb",
+							},
+						},
+					}...)
+				}
 			}
-			// This looks like its out of its place as we would expect this modules to be in the initrd but this is for Trusted Boot
-			// so the initrd is creating during artifact build and contains the rootfs, so this is ok to be in here
-			stages = append(stages, []schema.Stage{
-				{
-					Name:     "Download linux-modules-extra for nvdimm modules",
-					OnlyIfOs: "Ubuntu.*",
-					Commands: []string{
-						fmt.Sprintf("apt-get download linux-modules-extra-%s", kernel),
-						fmt.Sprintf("dpkg-deb -x linux-modules-extra-%s_*.deb /tmp/modules", kernel),
-						fmt.Sprintf("mkdir -p /usr/lib/modules/%s/kernel/drivers/nvdimm", kernel),
-						fmt.Sprintf("mv /tmp/modules/lib/modules/%[1]s/kernel/drivers/nvdimm/* /usr/lib/modules/%[1]s/kernel/drivers/nvdimm/", kernel),
-						fmt.Sprintf("depmod -a %s", kernel),
-						"rm -rf /tmp/modules",
-						"rm /*.deb",
-					},
-				},
-			}...)
 		}
 	}
 
@@ -916,7 +923,7 @@ func GetKairosInitramfsFilesStage(sis values.System, l types.KairosLogger) ([]sc
 					},
 				},
 			},
-			// Ubuntu 20.04 does not support the dracut multipath module 
+			// Ubuntu 20.04 does not support the dracut multipath module
 			// therefore we don't support multipath for Ubuntu 20.04 and below
 			{
 				Name:     "Add Multipath module to initramfs for Ubuntu 21.04 and above",
